+++ CHANGELOG.md
@@ -5,6 +5,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2019-04-13
+
+### Added
+
+### Changed
+
+### Removed
+
+### Fixed
+
+- Fixed a crash while listing outdated dependencies.
+
+
+## [0.4.0] - 2019-04-13
+
+### Added
+
+- Added ability to update outdated dependencies, with `jetzt --update`. This will list all the outdated dependencies based on command `jetzt --outdated`.
+
+### Changed
+
+### Removed
+
+### Fixed
+
+- Listing of outdated dependencies now distinguishes `dependencies` from `dev_dependencies`.
+
+
 ## [0.3.1] - 2019-04-12
 
 ### Added